pyproject.toml
[project]
name = "onbaord"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.47.2",
    "boto3>=1.37.0",
    "fastapi>=0.115.8",
    "instructor>=1.7.2",
    "motor>=3.7.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.34.0",
]
